More preparation!
I know this isn't the most descriptive title, but it gives a good idea of what I'm working on currently. See, we don't have much of a plan for the new robot. So, I want to setup our code structure beforehand, so that we can more efficiently do everything else. For this reason, I've started trying to organize the code in a way that makes sense to me, though probably not the best way.