Superstructure Changes and the 50 Comments
Starting with the small thing: I added a bunch of comments to functions just to clarify what they do when it isn't completely obvious.

Bigger thing: I've thought more about how the superstructure should work. I would like it to have three separate state machines (which is a lot). It has a robot mode, which is just a description of what category of action the robot is doing. There is then a wanted state, which handles each button press, and then a robot state, which is the actual action the robot is taking.

A method handleStateTransition() is used to decide, based on the mode, wanted state and sensors, what robot state should be used. This is then triggering wanted states in other subsystems, which influence robot action. Triggers will likely be used for this last part.

Even bigger thing: There's a isDisabled trigger now. You're welcome.