Effector and Elevator!
We've decided on a basic archetype for our robot. It will have a ground intake, elevator, and a pivoting end effector. Each subsystem will have state machines. Until I know how each thing will work fully, I can't do much of this, but this is a start. I decided to wait to setup the ground intake, and will do this tomorrow.