Current Work Finished!!
The largest issue currently is an odd disconnect issue that the dashboard software might be causing. This happens even in simulation. I can't tell if this is a program issue or not. Also, our CTRE Pigeon 2 is drifting, which is a cause for concern. I want to add standard deviations to the vision system, so that might be a next addition.